fix etudiant controller json resposne update

+++ src/main/java/esprit/tn/springdemo/controllers/EtudiantController.java
@@ -1,8 +1,10 @@
 package esprit.tn.springdemo.controllers;
 
 import esprit.tn.springdemo.entities.Etudiant;
+import esprit.tn.springdemo.responses.ApiResponse;
 import esprit.tn.springdemo.services.IEtudiantService;
 import lombok.AllArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,28 +16,86 @@ public class EtudiantController {
     private final IEtudiantService iEtudiantService;
 
     @GetMapping("")
-    public List<Etudiant> retrieveAllEtudiants() {
-        return iEtudiantService.retrieveAllEtudiants();
+    public ResponseEntity<ApiResponse> getEtudiants() {
+        ApiResponse apiResponse = new ApiResponse();
+        try {
+            List<Etudiant> etudiants = iEtudiantService.retrieveAllEtudiants();
+            apiResponse.setResponse(org.springframework.http.HttpStatus.OK, "Etudiants retrieved");
+            apiResponse.addData("etudiants", etudiants);
+        } catch (Exception e) {
+            apiResponse.setResponse(org.springframework.http.HttpStatus.BAD_REQUEST, e.getMessage());
+        }
+        return new ResponseEntity<>(apiResponse, apiResponse._getHttpStatus());
     }
 
     @PostMapping("")
-    public List<Etudiant> addEtudiants(@RequestBody List<Etudiant> etudiants) {
-        return iEtudiantService.addEtudiants(etudiants);
+    public ResponseEntity<ApiResponse> addEtudiants(@RequestBody List<Etudiant> etudiants) {
+        ApiResponse apiResponse = new ApiResponse();
+        try {
+            List<Etudiant> addedEtudiants = iEtudiantService.addEtudiants(etudiants);
+            //throw new RuntimeException("Test exception");
+            apiResponse.setResponse(org.springframework.http.HttpStatus.CREATED, "Etudiants added");
+            apiResponse.addData("etudiants", addedEtudiants);
+        } catch (Exception ex) {
+            apiResponse.setResponse(org.springframework.http.HttpStatus.BAD_REQUEST, ex.getMessage());
+        }
+        return new ResponseEntity<>(apiResponse, apiResponse._getHttpStatus());
     }
 
-    @PutMapping("")
-    public Etudiant updateEtudiant(@RequestBody Etudiant e) {
-        return iEtudiantService.updateEtudiant(e);
+    @PutMapping("/{idEtudiant}")
+    public ResponseEntity<ApiResponse> updateEtudiant(@RequestBody Etudiant etudiant, @PathVariable long idEtudiant) {
+        ApiResponse apiResponse = new ApiResponse();
+        try {
+            Etudiant foundEtudiant = iEtudiantService.retrieveEtudiant(idEtudiant);
+            if (foundEtudiant == null) {
+                apiResponse.setResponse(org.springframework.http.HttpStatus.NOT_FOUND, "Etudiant not found");
+            } else {
+                etudiant.setIdEtudiant(idEtudiant);
+                Etudiant updatedEtudiant = iEtudiantService.updateEtudiant(etudiant);
+                apiResponse.setResponse(org.springframework.http.HttpStatus.OK, "Etudiant updated");
+                apiResponse.addData("etudiant", updatedEtudiant);
+            }
+
+        } catch (Exception e) {
+            apiResponse.setResponse(org.springframework.http.HttpStatus.BAD_REQUEST, e.getMessage());
+        }
+        return new ResponseEntity<>(apiResponse, apiResponse._getHttpStatus());
     }
 
     @GetMapping("/{idEtudiant}")
-    public Etudiant retrieveEtudiant(@PathVariable long idEtudiant) {
-        return iEtudiantService.retrieveEtudiant(idEtudiant);
+    public ResponseEntity<ApiResponse> getEtudiant(@PathVariable long idEtudiant) {
+        ApiResponse apiResponse = new ApiResponse();
+        try {
+            Etudiant etudiant = iEtudiantService.retrieveEtudiant(idEtudiant);
+            if (etudiant == null) {
+                apiResponse.setResponse(org.springframework.http.HttpStatus.NOT_FOUND, "Etudiant not found");
+            } else {
+                apiResponse.setResponse(org.springframework.http.HttpStatus.OK, "Etudiant retrieved");
+                apiResponse.addData("etudiant", etudiant);
+            }
+        } catch (Exception e) {
+            apiResponse.setResponse(org.springframework.http.HttpStatus.BAD_REQUEST, e.getMessage());
+        }
+        return new ResponseEntity<>(apiResponse, apiResponse._getHttpStatus());
     }
 
     @DeleteMapping("/{idEtudiant}")
-    public void removeEtudiant(@PathVariable long idEtudiant) {
-        iEtudiantService.removeEtudiant(idEtudiant);
+    public ResponseEntity<ApiResponse> deleteEtudiant(@PathVariable long idEtudiant) {
+        ApiResponse apiResponse = new ApiResponse();
+        try {
+            Etudiant etudiant = iEtudiantService.retrieveEtudiant(idEtudiant);
+            if (etudiant == null) {
+                apiResponse.setResponse(org.springframework.http.HttpStatus.NOT_FOUND, "Etudiant not found");
+                return new ResponseEntity<>(apiResponse, apiResponse._getHttpStatus());
+            } else {
+                iEtudiantService.removeEtudiant(idEtudiant);
+                apiResponse.setResponse(org.springframework.http.HttpStatus.OK, "Etudiant deleted");
+                //apiResponse.addData("etudiant", etudiant);
+            }
+        } catch (Exception e) {
+            apiResponse.setResponse(org.springframework.http.HttpStatus.BAD_REQUEST, e.getMessage());
+        }
+        return new ResponseEntity<>(apiResponse, apiResponse._getHttpStatus());
     }
 
 }
